feat(UserService): Detalles despues de comments
